+++ CHANGELOG.md
@@ -13,6 +13,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Implemented `Arbitrary` for PHMM-related types under the `fuzzing feature
 - Added `SamHmmWriter` for writing pHMMs to SAM model files
 - Added `expanded_cigar_iter` for iterating over the operations in a CIGAR string
+- Added `IsValidDNA` strategies for non-ambiguous bases
 
 ### Changed